+++ Source/LocalizeStatic.swift
@@ -14,8 +14,8 @@ extension Localize {
     /// That prevent replace untagged values
     ///
     /// - returns: localized key or same text
-    public static func localize(key:String) -> String {
-        return Localize.shared.localize(key: key)
+    public static func localize(key:String, tableName:String? = nil) -> String {
+        return Localize.shared.localize(key: key, tableName: tableName)
     }
     
     /// Localize a string using your JSON File
@@ -24,8 +24,8 @@ extension Localize {
     /// - parameter value: The replacement value
     ///
     /// - returns: localized key or same text
-    public static func localize(key:String, replace:String) -> String {
-        return Localize.shared.localize(key: key, replace: replace)
+    public static func localize(key:String, replace:String, tableName:String? = nil) -> String {
+        return Localize.shared.localize(key: key, replace: replace, tableName: tableName)
     }
     
     /// Localize a string using your JSON File
@@ -34,8 +34,8 @@ extension Localize {
     /// - parameter value: The replacement values
     ///
     /// - returns: localized key or same text
-    public static func localize(key:String, values replace:[Any]) -> String {
-        return Localize.shared.localize(key: key, values: replace)
+    public static func localize(key:String, values replace:[Any], tableName:String? = nil) -> String {
+        return Localize.shared.localize(key: key, values: replace, tableName: tableName)
     }
     
     /// Localize string with dictionary values
@@ -45,8 +45,8 @@ extension Localize {
     /// - parameter value: The replacement dictionary
     ///
     /// - returns: localized key or same text
-    public static func localize(key:String, dictionary replace:[String:String]) -> String {
-        return Localize.shared.localize(key: key, dictionary: replace)
+    public static func localize(key:String, dictionary replace:[String:String], tableName:String? = nil) -> String {
+        return Localize.shared.localize(key: key, dictionary: replace, tableName: tableName)
     }
     
     /// Return storaged language or default language in device
@@ -99,6 +99,13 @@ extension Localize {
     /// Please not use in your code, is only for test schema.
     public static func testingMode() {
         Localize.shared.testingMode()
+    }
+    
+    // MARK: Config providers
+    
+    /// Update provider to localize your app.
+    public static func update(provider: LocalizeType) {
+        Localize.shared.update(provider: provider)
     }
     
 }