+++ Tests/ReadingOtherFiles.swift
@@ -12,10 +12,7 @@ class ReadingOtherFiles: XCTestCase {
     
     override func setUp() {
         super.setUp()
-    }
-    
-    override func tearDown() {
-        super.tearDown()
+        Localize.update(provider: .json)
     }
     
     func testOtherEnglish() {
@@ -40,13 +37,6 @@ class ReadingOtherFiles: XCTestCase {
         Localize.update(fileName: "file")
         let result = "hello".localize()
         XCTAssert(result == "Hallo Welt!")
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
     }
     
 }